+++ Assets/Scripts/TutorialScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7c1f2bc36cf1cdb4ab484b2389bea05d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
