+++ Assets/Sprites/freccia.png
